Implementing game settings loading
Implementing a settings config loading system to dynamically create
bingo games from configuration files

+++ Bingo Server/src/me/dylanmullen/bingo/game/currency/CurrencyController.java
@@ -28,7 +28,7 @@ public class CurrencyController
 		updateMySQL(u.getUUID(), u.getUserInformation().getCredits());
 	}
 
-	public void deduct(User u, int credits) throws InvalidAmountException
+	public void deduct(User u, double credits) throws InvalidAmountException
 	{
 		double userCredits = u.getUserInformation().getCredits() - credits;
 		if (userCredits < 0)
